requirements.txt
asgiref==3.8.1
black==24.4.0
Django==5.0.4
django-rest-framework==0.1.0
flake8==7.0.0
djangorestframework==3.15.1
mypy==1.9.0
pip==24.0
setuptools==65.5.0
sqlparse==0.5.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
pytz==2024.1
